refactor leadership
